Lazo meter
